presigned uploadテストのヘルパー関数を追加
