+++ apps/kaede/tests/storage/presigned-upload.test.ts
@@ -1,37 +1,11 @@
-import { GetObjectCommand, S3Client } from '@aws-sdk/client-s3'
 import { afterAll, beforeEach, describe, expect, it } from 'vitest'
 import {
   issueRecordingUploadUrls,
   resetS3ClientForTests,
 } from '../../src/services/storage/index.js'
+import { createStorageClient, readObjectText } from './support/helpers.js'
 
-const getRequiredEnv = (name: string) => {
-  const value = process.env[name]
-  if (!value) {
-    throw new Error(`${name} is not set`)
-  }
-
-  return value
-}
-
-const s3 = new S3Client({
-  region: getRequiredEnv('S3_REGION'),
-  endpoint: process.env.S3_INTERNAL_ENDPOINT,
-  credentials: {
-    accessKeyId: getRequiredEnv('S3_ACCESS_KEY_ID'),
-    secretAccessKey: getRequiredEnv('S3_SECRET_ACCESS_KEY'),
-  },
-  forcePathStyle: true,
-})
-
-const readBody = async (body: AsyncIterable<Uint8Array>) => {
-  const chunks: Uint8Array[] = []
-  for await (const chunk of body) {
-    chunks.push(chunk)
-  }
-
-  return Buffer.concat(chunks).toString('utf8')
-}
+const s3 = createStorageClient()
 
 describe('presigned upload integration', () => {
   beforeEach(() => {
@@ -62,20 +36,6 @@ describe('presigned upload integration', () => {
     })
 
     expect(uploadResponse.ok).toBe(true)
-
-    const object = await s3.send(
-      new GetObjectCommand({
-        Bucket: process.env.S3_BUCKET,
-        Key: `recordings/${recordingId}/raw/acce.csv`,
-      })
-    )
-    const body = object.Body
-
-    expect(body).toBeDefined()
-    if (!body || !(Symbol.asyncIterator in body)) {
-      throw new Error('response body is not async iterable')
-    }
-
-    await expect(readBody(body as AsyncIterable<Uint8Array>)).resolves.toBe(csv)
+    await expect(readObjectText(s3, `recordings/${recordingId}/raw/acce.csv`)).resolves.toBe(csv)
   })
 })
